backend/requirements.txt
# Django and DRF
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1

# API Documentation
drf-yasg==1.21.7

# JWT Authentication
djangorestframework-simplejwt==5.3.1

# Channels for WebSockets (optional)
channels==4.0.0
daphne==4.0.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Environment variables
python-dotenv==1.0.1

# Pillow for image processing
Pillow==10.2.0

# AWS S3 for file storage (optional)
django-storages==1.14.2
boto3==1.34.44

# Development tools
pytest==7.4.4
pytest-django==4.8.0
coverage==7.4.1
flake8==7.0.0
black==23.12.1

# WSGI server
gunicorn==21.2.0

# Dependencies
pytz==2024.1